\name{Biochemical_Oxygen_Demand}
\alias{Biochemical_Oxygen_Demand}
\docType{data}
\title{
Biochemical_Oxygen_Demand
}
\description{
Data set of time against oxygen usage for underwater microbes from water purity tests
}
\usage{data("Biochemical_Oxygen_Demand")}
\format{
  Data has the following columns
  \describe{
    \item{\code{Time}}{A numeric vector giving the time of the measurement (days)}
    \item{\code{Demand}}{A numeric vector giving the biochemical oxygen demand (mg/l).}
  }
}
\details{
}
\source{
Marske (1967), Biochemical Oxygen Demand Data Interpretation Using Sum of Squares Surface M.Sc. Thesis, University of Wisconsin – Madison.
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Biochemical_Oxygen_Demand)
## maybe str(Biochemical_Oxygen_Demand) ; plot(Biochemical_Oxygen_Demand) ...
}
\keyword{datasets}
