\name{Biomass_of_Herbiverous_Fish}
\alias{Biomass_of_Herbiverous_Fish}
\docType{data}
\title{
Biomass_of_Herbiverous_Fish
}
\description{
Biomass of deep and shallow Abobrinha Reef fish
}
\usage{data("Biomass_of_Herbiverous_Fish")}
\format{
  Data has the following colums
  \describe{
    \item{\code{Family}}{a string of fish familia}
    \item{\code{Species}}{a string of fish species}
    \item{\code{Morpho Functional Group}}{a string of feeding type}
    \item{\code{D.mean}}{a numeric vector of deep mean biomass}
    \item{\code{D.s.d.}}{a numeric vector of deep standard deviation}
    \item{\code{S.mean}}{a numeric vector of shallow mean biomass}
    \item{\code{S.s.d.}}{a numeric vector of shallow standard deviation}
  }
}
\details{
}
\source{
https://doi.org/10.3389/fmars.2020.00488.s001
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Biomass_of_Herbiverous_Fish)
## maybe str(Biomass_of_Herbiverous_Fish) ; plot(Biomass_of_Herbiverous_Fish) ...
}
\keyword{datasets}
